Added some Structural Elements
Added some structural elements to the README including horizontal line breaks to seperate the different headings in the document.

A link to the CS 305 Team DOS implementation of Gradebook was added and a comment was added to share the link to the Team NERDS implementation of Gradebook since those will be the base of the implementation for this project.