Create BranchStrategy.md
Created the markdown version of the Teams Branching Strategy. This is nearly an identical verision of the strategy with a few re-wordings here and there.

The Reference to Gradebook Requirements was turned into a live link to that document so that a user can click on the link to view the document. Anytime a branch is mentioned after the intial mention at the top of the document will be done in a code format (like `master`)