Added Team DOSs user Authentication
Added the User login information to the repo from Team DOSs repo.

This adds the functionality of allowing each user to have their own username and password to log into the system. This removes the need for the shared username and password that would be sent to the database when a user logs in.

This also adds the functionality of making all users fully accountable for their actions when using the system.

fixes #18