Fixed Format and Comments, Added Course link
Fixed some formatting issues where indentation was incorrect. Comments were added at the end of all major </div> tags so that it is easier to know what section is above where you are working. The course link was created so it will work when the div containing the ID for courses is added