Create addCourseMgmt.sql
Added the file addCourseMgmt.sql to the src/db folder. As of right now, the file only contains code for the 'insertCourse' function.

This function, when given the right parameters, will insert a row into the Course table.

It has not been tested yet and will need to undergo testing.