Added drop down selections
Added the drop down menus to the courses div, changed the id attributes for the selection boxes to have "reg" before them because of the need for the registrar to have different selections than an instructor would have.  Also added some comments to make it easier to understand what is going on in this section of the code.