Updated Course Mgmt Structure and Added Remove Course
Updated the structure of Add Course so that all of the input fields are now in a drop down menu. Added the Remove Course UI section which is also located in a drop down menu to save screen space if it is not needed. Remove a course contains 2 input fields, one for Number and one for Title along with a Remove Course button to submit the changes.