Added Basic add/remove UI for Section Mgmt
Added the section management UI for the HTML section of the code. Added the Add a section drop down menu with all of the drop down menus and text input felds to add a section. Also added the drop down for remove a section and the information needed to remove a section from the DB