Fixed Labels on add/remove course
Fixed the issue where the labels for the parameters of add and remove a course will not move to be above the input feld or will not stay above the input field when placeholder text is present. Also added the type parameter to the input tags to specify the type of input expected.

fixes #73