Retitled Search Button
Changed the title of the Search for courses button from Search for courses to Perform Search for ease of reading and to take up less screen space. NOTE: This will not effect the ID of the button, it only changes what the end user sees when looking at the button.