Create System Testing dir and Doc
Created a directory for system testing and a markdown document for the System Testing documentation that will be created for Gradebook. Added in header comment with some fields that may or may not change as time continues. Added in title, document purpose and created heading for Extra setup steps needed.