Added Course/Section Select and Fixed Label Issues
Added course and section select to the assign instructors drop down box so that all the needed informtion should be present now for this functionality. Remove instructors will be added shortly. Also created a fix for the input fields in add section where the input type="date" had the labels not moving when there is text in the field.