Added Remove Instructor UI
Added in the drop down menu for the remove instructor from a section functionality so that a dynamically populated table containing all of the instructors that are assigned to the section where the end user would have the ability to delete an instructor from the section as long as they are not the primary instructor of the section. The drop down also has 2 select menus for choosing the specific course and section that they want to remove the instructor from.