Added System Testing and underlines
Added in the section on System testing on the file structure section of the README describing what the System Testing portion of the file structure will be like. Also added was some underlining of the different important directories that have information following them to make it so they stand out more.